+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Allow "X" as final character of ORCID.
 - Ping the respective services to confirm the ORCIDs, RR IDs, and Uniprot IDs are actually good.
 - Add encoding as CLI param.
+- Add --offline option.
 
 ## v0.0.7 - 2021-01-13
 - Improved error messages in Excel.